Extract Store Results into results.rs.
We now have three files containing the different input/output types
associated with the store:

1. `types.rs` - shared domain types
2. `options.rs` - input options types
3. `results.rs` - output type

This is a nice separation, considering all this was previously in the
huge store.rs file, but store.rs is still extremely long (11K lines,
although ~6K of that is actually tests).
