+++ src/store/mod.rs
@@ -3,6 +3,7 @@
 
 mod options;
 mod ready_index;
+mod results;
 mod scheduled_index;
 mod store;
 mod types;
@@ -12,15 +13,16 @@ pub use options::{
     ListJobsOptions, PatchJobOptions, RetentionConfigPatch,
 };
 
+pub use results::{BulkCompleteResult, EnqueueResult, ListErrorsPage, ListJobsPage};
+
 pub use store::{
-    BulkCompleteResult, CommitMode, DEFAULT_BACKOFF_BASE_MS, DEFAULT_BACKOFF_EXPONENT,
-    DEFAULT_BACKOFF_JITTER_MS, DEFAULT_CACHE_SIZE, DEFAULT_COMPLETED_RETENTION_MS,
-    DEFAULT_DATA_TABLE_SIZE, DEFAULT_DEAD_RETENTION_MS, DEFAULT_INDEX_TABLE_SIZE,
-    DEFAULT_JOURNAL_SIZE, DEFAULT_L0_THRESHOLD, DEFAULT_RETRY_LIMIT, ListErrorsPage, ListJobsPage,
-    StorageConfig, Store, StoreEvent,
+    CommitMode, DEFAULT_BACKOFF_BASE_MS, DEFAULT_BACKOFF_EXPONENT, DEFAULT_BACKOFF_JITTER_MS,
+    DEFAULT_CACHE_SIZE, DEFAULT_COMPLETED_RETENTION_MS, DEFAULT_DATA_TABLE_SIZE,
+    DEFAULT_DEAD_RETENTION_MS, DEFAULT_INDEX_TABLE_SIZE, DEFAULT_JOURNAL_SIZE,
+    DEFAULT_L0_THRESHOLD, DEFAULT_RETRY_LIMIT, StorageConfig, Store, StoreEvent,
 };
 
 pub use types::{
-    BackoffConfig, EnqueueResult, EnvConfigError, ErrorRecord, Job, JobStatus, RetentionConfig,
-    ScanDirection, StoreError, UniqueConstraint, UniqueWhile,
+    BackoffConfig, EnvConfigError, ErrorRecord, Job, JobStatus, RetentionConfig, ScanDirection,
+    StoreError, UniqueConstraint, UniqueWhile,
 };
